\name{lsvsd-package}
\alias{lsvsd-package}
\docType{package}
\title{
Variable Selection Deviation
}
\description{
This package calculate the variable selection deviation (VSD) to measure the reliability of the given model. 
}
\details{
\tabular{ll}{
Package: \tab lsvsd\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-12\cr
License: \tab What license is it under?\cr
}
The package is very simple to use. It calculates the variable selection deviation of the model selected. The functions: \code{lsvsd,ARMweight,ARMPweight,AICweight,AICPweight,BICweight,BICPweight}
}
\author{
Ying Nan and Yuhong Yang

Maintainer: Ying Nan <nanx0006@umn.edu>

}
\references{
Ying Nan and Yuhong Yang (2012) VARIABLE SELECTION DIAGNOSTICS MEASURES FOR HIGH-DIMENSIONAL REGRESSION
}
